refactor(mika#1448): extract task_state/ module from db.rs
Move Task struct, related types (NewTask, ForcePromoteResult,
BackgroundTaskCounts, OrphanedParentTask, CompletableParentTask,
ReaperChildSnapshot, TaskHealthAnomaly, TaskHealthSummary), and
constants (TASK_TYPE_*, VALID_TASK_TYPES) into a new
crate::task_state::tasks submodule. Rename task_metadata.rs to
task_state/metadata.rs. Re-export everything from task_state/mod.rs
for backward compatibility.

This is the first extraction unit of the task_state module plan —
types and constants only, no query/write method migration yet.
